Add CI pipeline and make the clippy gate real (Phase 10.1)
There was no CI. That is why `npm run build` was broken on `main` without
anyone noticing, and why the ROADMAP claimed it passed. This adds the gate.

CI (.github/workflows/ci.yml), on every push to main and every PR:
- frontend: npm ci, npm run build (tsc + vite), npm test
- rust (windows-latest): cargo fmt --check,
  cargo clippy --no-default-features --all-targets -- -D warnings,
  cargo test --no-default-features

The Rust job verifies the feature-off build on purpose: the default
`llama-cpp` feature needs LLVM/libclang plus a full llama.cpp native build and
covers no code path that ships today.

To make `-D warnings` enforceable rather than aspirational, this also fixes
the 35 existing lint findings and reformats with rustfmt:
- unused imports and variables; unreachable code in the feature-gated local
  inference paths, restructured so the disabled path returns early
- redundant casts, redundant clones in tests, needless as_bytes().len()
- derivable Default impls (ContextCompressionMode via #[default]),
  missing Default for AgentOrchestrator / AgentStateManager
- sort_by -> sort_by_key, map_or(true, ..) -> is_none_or
- ModelType::to_string -> impl Display
- a genuinely redundant `if` in the SSE line-splitting loop where both
  branches were identical

Two crate-level allows remain, each with a written reason in lib.rs:
too_many_arguments (the 8-11 parameter orchestration entry points; collapsing
them into request structs is tracked as Phase 10.9) and
should_implement_trait. await_holding_lock is allowed only on the two test
modules that serialize env-var mutation, so production instances still fail.

Also:
- .gitattributes pins src-tauri/gen/schemas/*.json to LF, so a Windows build
  stops producing a CRLF-only diff on every cargo check.
- ROADMAP: fixed the phase numbering, which was off by one from Phase 10
  onward (Phase 10 items were numbered 9.x and collided with real Phase 9
  items), and reprioritized Next Immediate Tasks. The permission model V2 is
  now item 1: MCP tool calls currently run without per-call approval.

+++ src-tauri/src/commands/git.rs
@@ -61,8 +61,8 @@ pub fn git_status(path: String) -> Result<GitStatus, String> {
         let (a, b) = repo
             .graph_ahead_behind(local.id(), upstream_commit.id())
             .map_err(|e| e.to_string())?;
-        ahead = a as usize;
-        behind = b as usize;
+        ahead = a;
+        behind = b;
     }
 
     // 获取状态
